Move exchanges api to clients folder
