+++ clients/coinbase/.gitignore
