+++ clients/coinbase/.travis.yml
