+++ clients/coinbase/LICENSE
