+++ clients/coinbase/client.go
