+++ clients/coinbase/client_test.go
