+++ clients/coinbase/currency.go
