+++ clients/coinbase/cursor.go
