+++ clients/coinbase/error.go
