+++ clients/coinbase/message.go
