+++ clients/coinbase/product.go
