+++ clients/coinbase/product_test.go
