+++ clients/coinbase/report.go
