+++ clients/coinbase/report_test.go
