+++ clients/coinbase/transfer_test.go
