+++ clients/gemini/.gitignore
