+++ clients/gemini/README.md
