+++ clients/gemini/gemini_util.go
