+++ clients/gemini/private.go
