+++ clients/gemini/public.go
